+++ src/libraries/System.Threading.Channels/ref/System.Threading.Channels.netcoreapp.cs
@@ -14,8 +14,4 @@ namespace System.Threading.Channels
 #endif
         protected ChannelClosedException(System.Runtime.Serialization.SerializationInfo info, System.Runtime.Serialization.StreamingContext context) { }
     }
-    public abstract partial class ChannelReader<T>
-    {
-        public virtual System.Collections.Generic.IAsyncEnumerable<T> ReadAllAsync(System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) { throw null; }
-    }
 }
