+++ src/libraries/System.Threading.Channels/tests/System.Threading.Channels.Tests.csproj
@@ -6,18 +6,16 @@
   <ItemGroup>
     <Compile Include="BoundedChannelTests.cs" />
     <Compile Include="ChannelClosedExceptionTests.cs" />
+    <Compile Include="ChannelClosedExceptionTests.netcoreapp.cs"
+             Condition="'$(TargetFramework)' == '$(NetCoreAppCurrent)'" />
     <Compile Include="ChannelTestBase.cs" />
     <Compile Include="ChannelTests.cs" />
+    <Compile Include="DebugAttributeTests.cs" />
+    <Compile Include="Stress.cs" />
     <Compile Include="TestBase.cs" />
     <Compile Include="UnboundedChannelTests.cs" />
-    <Compile Include="DebugAttributeTests.cs" />
     <Compile Include="$(CommonTestPath)System\Diagnostics\DebuggerAttributes.cs"
              Link="Common\System\Diagnostics\DebuggerAttributes.cs" />
-  </ItemGroup>
-  <ItemGroup Condition="'$(TargetFramework)' == '$(NetCoreAppCurrent)'">
-    <Compile Include="Stress.cs" />
-    <Compile Include="ChannelClosedExceptionTests.netcoreapp.cs" />
-    <Compile Include="ChannelTestBase.netcoreapp.cs" />
   </ItemGroup>
   <ItemGroup Condition="'$(TargetFramework)' == '$(NetFrameworkMinimum)'">
     <ProjectReference Include="..\src\System.Threading.Channels.csproj" />
